requirements.txt
numpy
pandas
pyarrow
matplotlib
jupyter
requests
python-dateutil
